# Optimize performance of depended_by_gems view by replacing improper join with subquery